Update README.md
This is for people trying to start learning about different Meanings and backgrounds of computers. This was made by Javon Pickett. I feel Like I gave a little facts on different types of words and different aspects that you can during this course. Digital information was a great section for me that we studied.